packages/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
httpx<0.25
web3
eth-account
celery
redis
pytest
